+++ Embedded/Editor/com.unity.cloud.common/Runtime/InternalsVisibleToAM4U.cs
@@ -6,5 +6,3 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.Cloud.Identity.Runtime.Embedded")]
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Editor.Embedded")]
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Runtime.Embedded")]
-[assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Tests.Editor.Embedded")]
-[assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Tests.Embedded")]